feat(contexts/auth): переименование файла по названию компонента и добавление функции login для поддержания single responsibility
