+++ frontend/src/contexts/authorization/types.ts
@@ -1,8 +1,8 @@
 export interface AuthContextType {
     isAuthenticated: boolean;
-    token: (accessToken: string) => void;
+    token: string;
     username: string;
     setUsername: (username: string) => void;
+    login: (token: string, username: string) => void;
     logout: () => void;
-    isLoading: boolean;
 }
